Finish package signing.
Package signing in Necessitas become a walk in the park.
Now Necessitas users, can enjoy a first class very easy to use and very
intuitive user interface to create/choose keys to sign their
applications in order to publish them on the market.

Fix "Cannot create children for a parent that is in a different thread".

DIGEST: